Major restructuring of program:
Earlier all functions were contained in the pieri_functions module, which also
utilized the equation_count.py module.  These files are no longer needed.
The program now consists of the following modules:
permutation_functions, schubert, k_mult, dual_classes, lenart_functions,
and module_homomorphism.

These modules are structured as follows:

-permutation_functions.py contains basic functions for working with signed
and unsigned permutations.

-schubert.py defines the Grassmannian class
and the various methods and object variables associated to it.  it depends on
permutation_functions.py.

-k_mult.py contains the necessary functions for calculating the product of
an arbitrary schubert class with a special schubert class.  most of the methods
take a Grassmannian object as one of the arguments, but it does not explicitly
refer to the schubert module.

-dual_classes.py contains methods for investigating the combinatorics of
dual classes, and in particular constructs the dual class of an arbirtary schubert class
using only local information.  it depends on permutation_functions.py, and most of its
methods take a Grassmannian object as an argument.

-lenart_functions.py implements Lenart and Postnikov's K-theoretic Monk's formula,
found in their paper 'Affine Weyl Groups in K-theory and Reprensentation Theory.'
It refers to permutation_functions.py and k_mult.py, and the methods take Grassmannian
objects as arguments.

-module_homomorphism.py investigates the possibility of a k-theoretic module homomorphism
 analogous to the one in equivariant cohomology discovered by Buch, Kresch, and Tamvakis,
in their paper 'A Giambelli formula for even orthogonal Grassmannians.'  It refers to
k_mult.py and the methods take Grassmannian objects as arguments.